♻️ refactor: 계정 설정 재사용 가능한 컴포넌트로 리팩터링
