--- apps/web/src/app/(needLogin)/mypage/setting/account/_components/AccountPrivateOrPublic/index.ts
@@ -1,1 +1,0 @@
-export * from './AccountPrivateOrPublic';
