+++ apps/web/src/app/(needLogin)/mypage/setting/account/_components/ModifyForm/ModifyForm.types.tsx
@@ -1,0 +1,5 @@
+export type ModifyType = {
+  password?: string;
+  pwconfirm?: string;
+  nickName?: string;
+};
