--- apps/web/src/app/(needLogin)/mypage/setting/account/_components/PWModifyForm/PWModifyForm.types.ts
@@ -1,4 +1,0 @@
-export type PWModifyType = {
-  password?: string;
-  pwconfirm?: string;
-};
