--- apps/web/src/app/(needLogin)/mypage/setting/account/_components/PWModifyForm/index.ts
@@ -1,2 +1,0 @@
-export * from './PWModifyForm';
-export * from './PWModifyForm.types';
