--- apps/web/src/app/(needLogin)/mypage/setting/account/_components/PrivacyModifyForm/PrivacyModifyForm.types.ts
@@ -1,3 +1,0 @@
-export type PrivacyModifyType = {
-  nickName?: string;
-};
