--- apps/web/src/app/(needLogin)/mypage/setting/account/_components/PrivacyModifyForm/index.ts
@@ -1,2 +1,0 @@
-export * from './PrivacyModifyForm';
-export * from './PrivacyModifyForm.types';
