--- apps/web/src/app/(needLogin)/mypage/setting/account/_components/index.ts
@@ -1,3 +1,0 @@
-export * from './AccountPrivateOrPublic';
-export * from './PrivacyModifyForm';
-export * from './PWModifyForm';
